requirements.txt
aerich==0.7.2
aiosmtplib==2.0.2
aiofiles==24.1.0
asyncpg==0.29.0
fastapi==0.115.7
fastapi-injector @ git+https://github.com/10XScale-in/fastapi-injector.git@8aee4c4b1c6b2db6c262eb5a86acb35a4d0d178d
firebase-admin==6.5.0
gunicorn==23.0.0
httpx==0.27.0
injector==0.22.0
lib==4.0.0
markdown-it-py==3.0.0
mkdocs-gen-files==0.5.0
mkdocstrings==0.25.2
mypy-extensions==1.0.0
oauth2client==4.1.3
orjson==3.10.6
pre-commit==3.8.0
python-multipart==0.0.19
pydantic==2.8.2
pydantic-settings==2.3.4
pytest==8.3.2
pytest-env==1.1.3
redis==5.0.7
requests==2.32.3
ruff==0.5.2
sentry-sdk==2.10.0
snowflakeid @ git+https://github.com/10XScale-in/snowflakeid.git
starlette==0.40.0
strawberry-graphql==0.257.0
taskiq==0.11.10
taskiq-redis==1.0.0
tortoise==0.1.1
tortoise-orm==0.22.1
uvicorn==0.30.1